Extract an Evaluation Curve

Extract an Evaluation Curve

Usage

extractEvalCurve(
  evalrec,
  xlim = NULL,
  as.list = FALSE,
  resolution = NULL,
  sig.scale = NULL
)

Arguments

evalrec

data.frame. A single evaluation record from NASIS_evaluations data set.

xlim

numeric. Length 2. Minimum and maximum values to override internal domain range.

as.list

Return list with function and domain limits? Default FALSE returns just the function.

resolution

Deprecated. Not used.

sig.scale

Deprecated. Not used.

Value

function that can evaluate input data and return evaluation values. When as.list=TRUE a list object containing elements FUN (the function) and domain (a numeric vector containing domain minimum and maximum)